requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
python-dotenv
requests
pydantic